Added ability to perform a stall and improved scenario generation.
Stall, isn't used currently, but exists as a feature we could try to
add.  Allows the path finder to purposely waste time, with the hope it
could lead to a better solution.  Needs to be deployed in the right
circumstances.

Added ability to iteratively generate scenarios.  Code needs to be
cleaned up/improved.